fix lora patch
